rk3399: clock: add reset interface for PCIe
w/o the basic framework of reset controller, it becomes
difficult for us to consolidate the reset ops inside the
box. It is also pointless to introduce dt-bindings parse
helper function to get them. Hardcode it in the rk3399
clock file and expose the API which keeps the model as-is.

Change-Id: Ie0a30b6642d410c28bad56413134b7099e0fe1c6
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Signed-off-by: Bin Yang <yangbin@rock-chips.com>
(cherry picked from commit 78dd11d2ca6e04605296e19c168ad97f0ad53af9)