BACKPORT: usb: xhci: Guard all calls to xhci_wait_for_event
xhci_wait_for_event returns NULL on timeout, so the caller always has to
check for that. This addresses immediate explosions in this part
of the code when timeouts happen, but not the root cause for the
timeout.

Change-Id: Ib899b615349ed5d1d00885b6508c765b6e773d6a
Signed-off-by: Hector Martin <marcan@marcan.st>
Reviewed-by: Marek Vasut <marex@denx.de>
Signed-off-by: William Wu <william.wu@rock-chips.com>
(cherry picked from commit 8d1e03f984c7467d7c8883f15dea14b2f8b4c0e2)