rockchip: rk3576: Enable cci channels for devices
CCI channel is a path for AXI to get data from CPU cache if available
instead of from DRAM.

This is a hardware cache coherent feature for device access to memory.
With this feature enabled, and "dma-coherent" in dts node, the modules
in kernel can use cacheable memory without mantain cache coherent manually.

No kernel driver needs to change, the memory system can handle it with
"dma-coherent" property in dts.

Change-Id: Ibd30ab34e7fe1de1827282634355d73adc4b0bbc
Signed-off-by: Kever Yang <kever.yang@rock-chips.com>