video/drm: dw_dp: create a device for dp port
when DPTX support mst, it will mulit port in dts as follow:
dp {
	dp0 {
		...
	};
	dp1 {
		...
	};
	...
};

The display route is described in the child node and not
the top device node. It's better to create a device for
the child node(dp0 node) for the rochchip display drvier
to parse the display route correctly.

Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>
Change-Id: Ibf02e172eac22de83d86dcb7757fd162198234a7