phy: rockchip: naneng-combphy: fix phy device mismatch
The uclass_get_device_by_driver() API always finds the first device in
DT probed, however, it may not the PHY that associate with USB download
controller, if mismathed, the below errors would be printed in terminal.

  download key pressed... entering download mode...
  RKUSB: LUN 0, dev 0, hwpart 0, sector 0x0, count 0x1d5a000
  failed to enable ep0out
  failed to start <NULL>: -110
  g_dnl_register: failed!, error: -110
  g_dnl_register failedEnter rockusb failed, fallback to bootrom...

So passed the PHY physical address and loop to match the correct PHY
device to fix it.

Fixes: 14d5da7dccb8 ("phy: rockchip: naneng-combphy: Add usb3 phy init for rockusb")
Change-Id: Ib65a104008d6302dca09cc081a3457b3d55e0323
Signed-off-by: Frank Wang <frank.wang@rock-chips.com>