dm: regulator: Enable fixed reglator all the time
REGULATOR_FLAG_AUTOSET_UV will be set if min_uV == max_uV, and then
the regulator_set_value() will be called and returns -ENOSYS for
fixed regulator, it makes the following regulator_set_enable() was
not called.

This patch follow upstream solution, but we have to ignore result
to compabile the old possible failure before this patch.

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: I0d89c0abc48c29c913cf30ad4ae919852750c2b8