video/drm: analogix_dp: Fix incorrect usage of enhanced mode
Enhanced mode is required by the eDP 1.2 specification, and not doing it
early could result in a period of time where we have a link transmitting
idle packets without it. Since there is no reason to disable it, we just
enable it at the beginning of link training and then keep it on all the
time.

Related kernel commit:
1932250df1f0 ("drm/bridge: analogix_dp: Fix incorrect usage of enhanced mode")

Fixes: 186f85721af ("drm/rockhcip: add drm rockchip display support")
Change-Id: I32be24ae8561b79dce549e1fd55af9bd0341de2e
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>