video/drm: dsi2: disable BTA during auto calculation training
when DSI operates in auto calculation mode,it sends a dsc read
get_scanline request to the peripheral devices to automatically
calculate phy_max_rd_time_auto during the auto calculation training.
However,if the peripheral devices,such as a bridge chip or some panel,
lack the capability to respond to read-back requests,they cannot respond
to the DSI host’s BTA,leading to the signal control not being properly
returned to the DSI host,Therefore,the BTA function should be disabled
in such case.

Change-Id: If1f6925f2614fce557ef4200275b14879a29cbd1
Signed-off-by: Guochun Huang <hero.huang@rock-chips.com>