video/drm: vop2: fix the default plane mask configurations for RK3588
The relationships between the main window and the splice window are:

Main     | Splice
Cluster0 | Cluster1
Cluster2 | Cluster3
Esmart0  | Esmart1
Esmart2  | Esmart3

The VP0 and VP1 should be used in combination when the display mode is
over 4k. The main window should attach to the VP0, while the splice
window should attach to the other.

If only one VP is enabled and the plane mask is not assigned in DTS, all
main windows will be assigned to the enabled VPx, and all splice windows
will be assigned to the VPx+1, in order to ensure that the splice mode
work well.

Without this patch, the default plane_mask may not meet the above
requirement when the plane_mask is not assigned in DTS.

Change-Id: I313ba17f0ffad958c1bdb2e3caca74ba85dc5da9
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>