dm: serial: set flag GD_FLG_SERIAL_READY if CONFIG_ROCKCHIP_VIDCONSOLE is enabled
For RK platforms, the CONFIG_DEBUG_UART_ALWAYS is enabled by
default, so the GD_FLG_SERIAL_READY will not be set in serial_init()
of drivers/serial/serial_uclass.c, which makes stdout support only
"serial".

If GD_FLG_SERIAL_READY is not set, the console will only use
the function printch() to show the logs. The related codes in
common/console.c are:

void putc(const char c)
......
	/* if we don't have a console yet, use the debug UART */
	if (!gd || !(gd->flags & GD_FLG_SERIAL_READY)) {
		printch(c);
		return;
	}
......
}

In order to show logs on srcreen via the video console, we set
the flag GD_FLG_SERIAL_READY in serial_init() if the config
CONFIG_ROCKCHIP_VIDCONSOLE is enabled.

Change-Id: I4a10ac7a6b1f0798bcf4ebfef11a1214719fe6fc
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>