video/drm: vop2: update the calculation of dclk_out division factor for RK3588 HDMI DSC mode
For RK3588, dclk_out is designed for DP, MIPI in both DSC and non-DSC
mode and HDMI in DSC mode.

In addditon, the dclk_out division factor should be the same as the
dclk_core division factor in HDMI DSC mode.

For the display interfaces that do not need the dclk_out, it should
be more appropriate to set dclk_out division factor to 0.

Change-Id: Idd254a105b5ea69caa25b315be8ff39a8c4fd10e
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>